web/requirements.txt
requests==2.11.1
django-phonenumber-field==1.1.0
